# docs/frontend/openapi.yaml
openapi: 3.0.0
info:
  title: Respondr.ai Backend API
  description: Mass Casualty Triage & Logistics Assistant API
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Local Development Server
paths:
...
  /hospitals:
    get:
      summary: List Hospitals
      description: List all hospitals or filter by proximity.
      parameters:
        - name: lat
          in: query
          schema:
            type: number
          description: Latitude for proximity search
        - name: lon
          in: query
          schema:
            type: number
          description: Longitude for proximity search
        - name: max_km
          in: query
          schema:
            type: integer
            default: 10
          description: Maximum distance in kilometers
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Hospital'
  /reserve_bed:
    post:
      summary: Reserve Bed
      description: Reserve bed(s) at a specific hospital with safety buffer logic.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BedReservation'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationResponse'
        '404':
          description: Hospital not found
  /api/allocate:
    post:
      summary: Allocate Beds (Bulk)
      description: Main allocation endpoint - distribute patients across nearby hospitals.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AllocationRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AllocationPlan'
        '400':
          description: Bad Request (e.g., invalid coordinates)
...
components:
  schemas:
...
    Hospital:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        type:
          type: string
          default: Feature
        geometry:
          $ref: '#/components/schemas/GeoJSONGeometry'
        properties:
          $ref: '#/components/schemas/HospitalProperties'
        allocated:
          type: integer
          default: 0
        last_updated:
          type: string
          format: date-time
      required: [_id, name, geometry, properties]
    GeoJSONGeometry:
      type: object
      properties:
        type:
          type: string
          default: Point
        coordinates:
          type: array
          items:
            type: number
          minItems: 2
          maxItems: 2
          description: "[longitude, latitude]"
      required: [type, coordinates]
    HospitalProperties:
      type: object
      properties:
        zone:
          type: string
        icu_capacity:
          type: integer
        trauma_capacity:
          type: integer
        specialties:
          type: array
          items:
            type: string
        contact:
          type: string
      required: [zone, icu_capacity, trauma_capacity]
    BedReservation:
      type: object
      properties:
        hospital_id:
          type: string
        severity:
          type: string
          enum: [RED, YELLOW, GREEN, BLACK]
        patient_count:
          type: integer
          default: 1
      required: [hospital_id, severity]
    ReservationResponse:
      type: object
      properties:
        status:
          type: string
          enum: [SUCCESS, CONFLICT]
        hospital:
          type: string
        beds_reserved:
          type: integer
        remaining_capacity:
          type: integer
        available:
          type: integer
        message:
          type: string
      required: [status]
    AllocationRequest:
      type: object
      properties:
        lat:
          type: number
        lon:
          type: number
        patient_count:
          type: integer
        triage_color:
          type: string
          enum: [RED, YELLOW, GREEN, BLACK]
          default: RED
      required: [lat, lon, patient_count]
    AllocationPlan:
      type: object
      properties:
        status:
          type: string
          enum: [SUCCESS, PARTIAL, OVERFLOW]
        plan:
          type: array
          items:
            $ref: '#/components/schemas/HospitalAllocation'
        unallocated:
          type: integer
          default: 0
        message:
          type: string
      required: [status, plan]
    HospitalAllocation:
      type: object
      properties:
        hospital_id:
          type: string
        hospital_name:
          type: string
        patients_sent:
          type: integer
        coordinates:
          type: array
          items:
            type: number
        distance_km:
          type: number
      required: [hospital_id, hospital_name, patients_sent, coordinates, distance_km]
...